Weight odds to sporadic rather than regular
The seed drop from sharp stone use on grass is supposed to be a sporadic happening rather than an intentional strategy to get seeds.  I don't want this to be how people WANT to try to get seeds.  It's a byproduct.  Hoes are the actual way to try for seeds.